+++ tapir/core/tests/test_mailoption/test_MailSettingsView.py
@@ -32,27 +32,47 @@ class MandatoryMail(TapirEmailBase):
         return "MandatoryMail"
 
 
+def setup_mails():
+    TapirEmailBase.register_email(OptionalNonDefaultMail)
+    TapirEmailBase.register_email(MandatoryMail)
+
+
 class TestMailSetting(TapirFactoryTestBase):
     def setUp(self):
-        TapirEmailBase.register_email(OptionalNonDefaultMail)
-        TapirEmailBase.register_email(MandatoryMail)
+        setup_mails()
 
-    def test_userWantsToOrHasToReceiveMail_mailNotSubscribedTo_shouldNotReceiveMail(
-        self,
-    ):
-        assert OptionalNonDefaultMail.option == MailOption.OPTIONAL_DISABLED
-        tapir_user: TapirUser = TapirUserFactory.create()
+    def test_NormalUser_UpdateOwnMailSetting_returnsTrue(self):
+        tapir_user = TapirUserFactory()
+        self.login_as_user(tapir_user)
+
         self.assertFalse(
             OptionalNonDefaultMail().user_wants_to_or_has_to_receive_mail(
                 user=tapir_user
             )
         )
 
-    def test_userWantsToOrHasToReceiveMail_mailSubscribedTo_shouldReceiveMail(self):
-        tapir_user: TapirUser = TapirUserFactory.create()
-        OptionalMails.objects.create(
-            user=tapir_user, mail_id=OptionalNonDefaultMail.get_unique_id(), choice=True
+        response = self.send_request_add_OptionalNonDefaultMail_to_subscribed_mails(
+            tapir_user
         )
+        self.assertStatusCode(response=response, expected_status_code=HTTPStatus.OK)
+
+        tapir_user.refresh_from_db()
+        self.assertTrue(
+            OptionalNonDefaultMail().user_wants_to_or_has_to_receive_mail(
+                user=tapir_user
+            )
+        )
+
+    def test_MemberOffice_canUpdateOtherUsersMailSettings_returnsTrue(self):
+        tapir_user = TapirUserFactory()
+        self.login_as_member_office_user()
+
+        response = self.send_request_add_OptionalNonDefaultMail_to_subscribed_mails(
+            usr=tapir_user
+        )
+        self.assertStatusCode(response=response, expected_status_code=HTTPStatus.OK)
+
+        tapir_user.refresh_from_db()
         self.assertTrue(
             OptionalNonDefaultMail().user_wants_to_or_has_to_receive_mail(
                 user=tapir_user
@@ -105,60 +125,3 @@ class TestMailSetting(TapirFactoryTestBase):
             reverse("accounts:mail_settings", args=[tapir_user.pk]),
         )
         self.assertTrue(200, response.status_code)
-
-    def test_NormalUser_UpdateOwnMailSetting_returnsTrue(self):
-        tapir_user = TapirUserFactory()
-        self.login_as_user(tapir_user)
-
-        self.assertFalse(
-            OptionalNonDefaultMail().user_wants_to_or_has_to_receive_mail(
-                user=tapir_user
-            )
-        )
-
-        response = self.send_request_add_OptionalNonDefaultMail_to_subscribed_mails(
-            tapir_user
-        )
-        self.assertStatusCode(response=response, expected_status_code=HTTPStatus.OK)
-
-        tapir_user.refresh_from_db()
-        self.assertTrue(
-            OptionalNonDefaultMail().user_wants_to_or_has_to_receive_mail(
-                user=tapir_user
-            )
-        )
-
-    def test_MemberOffice_canUpdateOtherUsersMailSettings_returnsTrue(self):
-        tapir_user = TapirUserFactory()
-        self.login_as_member_office_user()
-
-        response = self.send_request_add_OptionalNonDefaultMail_to_subscribed_mails(
-            usr=tapir_user
-        )
-        self.assertStatusCode(response=response, expected_status_code=HTTPStatus.OK)
-
-        tapir_user.refresh_from_db()
-        self.assertTrue(
-            OptionalNonDefaultMail().user_wants_to_or_has_to_receive_mail(
-                user=tapir_user
-            )
-        )
-
-    def test_NormalUser_CannotUpdateNonOptionalMailSettings(self):
-        tapir_user: TapirUser = TapirUserFactory.create()
-        self.login_as_user(tapir_user)
-        # Enabled by default and mandatory
-        self.assertTrue(
-            MandatoryMail().user_wants_to_or_has_to_receive_mail(user=tapir_user)
-        )
-
-        post_data = {"mandatory_mails": []}
-        self.client.post(
-            reverse("accounts:mail_settings", args=[tapir_user.pk]),
-            post_data,
-            follow=True,
-        )
-        tapir_user.refresh_from_db()
-        self.assertTrue(
-            MandatoryMail().user_wants_to_or_has_to_receive_mail(user=tapir_user)
-        )
